Implement the new block subsidy function (#1830)
* Replace the old blockSubsidy parameters with the new ones.

* Return subsidyGenesisReward if blockHash is the genesis hash.

* Traverse a block's past for the subsidy calculation.

* Partially implement SubsidyStore.

* Refer to SubsidyStore from CoinbaseManager.

* Wrap calcBlockSubsidy in getBlockSubsidy, which first checks the database.

* Fix finalityStore not calling GenerateShardingID.

* Implement calculateAveragePastSubsidy.

* Implement calculateMergeSetSubsidySum.

* Implement calculateSubsidyRandomVariable.

* Implement calcBlockSubsidy.

* Add a TODO about floats.

* Update the calcBlockSubsidy TODO.

* Use binary.LittleEndian in calculateSubsidyRandomVariable.

* Fix bad range in calculateSubsidyRandomVariable.

* Replace float64 with big.Rat everywhere except for subsidyRandomVariable.

* Fix a nil dereference.

* Use a random walk to approximate the normal distribution.

* In order to avoid unsupported fractional results from powInt64, flip the numerator and the denominator manually.

* Set standardDeviation to 0.25, MaxSompi to 10_000_000_000 * SompiPerKaspa and defaultSubsidyGenesisReward to 1_000.

* Set the standard deviation to 0.2.

* Use a binomial distribution instead of trying to estimate the normal distribution.

* Change some values around.

* Clamp the block subsidy.

* Remove the fake duplicate constants in the util package.

* Reduce MaxSompi to only 100m Kaspa to avoid hitting the uint64 ceiling.

* Lower MaxSompi further to avoid new and exciting ways for the uint64 ceiling to be hit.

* Remove debug logs.

* Fix a couple of failing tests.

* Fix TestBlockWindow.

* Fix limitTransactionCount sometimes crashing on index-out-of-bounds.

* In TrustedDataDataDAABlock, replace BlockHeader with DomainBlock

* In calculateAveragePastSubsidy, use blockWindow instead of doing a BFS manually.

* Remove the reference to DAGTopologyManager in coinbaseManager.

* Add subsidy to the coinbase payload.

* Get rid of the subsidy store and extract subsidies out of coinbase transactions.

* Keep a blockWindow amount of blocks under the virtual for IBD purposes.

* Manually remove the virtual genesis from the merge set.

* Fix simnet genesis.

* Fix TestPruning.

* Fix TestCheckBlockIsNotPruned.

* Fix TestBlockWindow.

* Fix TestCalculateSignatureHashSchnorr.

* Fix TestCalculateSignatureHashECDSA.

* Fix serializing the wrong value into the coinbase payload.

* Rename coinbaseOutputForBlueBlock to coinbaseOutputAndSubsidyForBlueBlock.

* Add a TODO about optimizing trusted data DAA window blocks.

* Expand on a comment in TestCheckBlockIsNotPruned.

* In calcBlockSubsidy, divide the big.Int numerator by the big.Int denominator instead of converting to float64.

* Clarify a comment.

* Rename SubsidyMinGenesisReward to MinSubsidy.

* Properly handle trusted data blocks in calculateMergeSetSubsidySum.

* Use the first two bytes of the selected parent's hash for randomness instead of math/rand.

* Restore maxSompi to what it used to be.

* Fix TestPruning.

* Fix TestAmountCreation.

* Fix TestBlockWindow.

* Fix TestAmountUnitConversions.

* Increase the timeout in many-tips to 30 minutes.

* Check coinbase subsidy for every block

* Re-rename functions

* Use shift instead of powInt64 to determine subsidyRandom

+++ domain/consensus/processes/pruningmanager/pruningmanager.go
@@ -325,7 +325,11 @@ func (pm *pruningManager) deletePastBlocks(stagingArea *model.StagingArea, pruni
 		}
 	}
 
-	err = pm.deleteBlocksDownward(stagingArea, queue)
+	blocksToKeep, err := pm.calculateBlocksToKeep(stagingArea, pruningPoint)
+	if err != nil {
+		return err
+	}
+	err = pm.deleteBlocksDownward(stagingArea, queue, blocksToKeep)
 	if err != nil {
 		return err
 	}
@@ -333,9 +337,33 @@ func (pm *pruningManager) deletePastBlocks(stagingArea *model.StagingArea, pruni
 	return nil
 }
 
-func (pm *pruningManager) deleteBlocksDownward(stagingArea *model.StagingArea, queue model.BlockHeap) error {
+func (pm *pruningManager) calculateBlocksToKeep(stagingArea *model.StagingArea,
+	pruningPoint *externalapi.DomainHash) (map[externalapi.DomainHash]struct{}, error) {
+
+	pruningPointAnticone, err := pm.dagTraversalManager.AnticoneFromVirtualPOV(stagingArea, pruningPoint)
+	if err != nil {
+		return nil, err
+	}
+	pruningPointAndItsAnticone := append(pruningPointAnticone, pruningPoint)
+	blocksToKeep := make(map[externalapi.DomainHash]struct{})
+	for _, blockHash := range pruningPointAndItsAnticone {
+		blocksToKeep[*blockHash] = struct{}{}
+		blockWindow, err := pm.dagTraversalManager.BlockWindow(stagingArea, blockHash, pm.difficultyAdjustmentWindowSize)
+		if err != nil {
+			return nil, err
+		}
+		for _, windowBlockHash := range blockWindow {
+			blocksToKeep[*windowBlockHash] = struct{}{}
+		}
+	}
+	return blocksToKeep, nil
+}
+
+func (pm *pruningManager) deleteBlocksDownward(stagingArea *model.StagingArea,
+	queue model.BlockHeap, blocksToKeep map[externalapi.DomainHash]struct{}) error {
+
 	visited := map[externalapi.DomainHash]struct{}{}
-	// Prune everything in the queue including its past
+	// Prune everything in the queue including its past, unless it's in `blocksToKeep`
 	for queue.Len() > 0 {
 		current := queue.Pop()
 		if _, ok := visited[*current]; ok {
@@ -343,11 +371,16 @@ func (pm *pruningManager) deleteBlocksDownward(stagingArea *model.StagingArea, q
 		}
 		visited[*current] = struct{}{}
 
-		alreadyPruned, err := pm.deleteBlock(stagingArea, current)
-		if err != nil {
-			return err
+		shouldAddParents := true
+		if _, ok := blocksToKeep[*current]; !ok {
+			alreadyPruned, err := pm.deleteBlock(stagingArea, current)
+			if err != nil {
+				return err
+			}
+			shouldAddParents = !alreadyPruned
 		}
-		if !alreadyPruned {
+
+		if shouldAddParents {
 			parents, err := pm.dagTopologyManager.Parents(stagingArea, current)
 			if err != nil {
 				return err
@@ -927,13 +960,13 @@ func (pm *pruningManager) blockWithTrustedData(stagingArea *model.StagingArea, b
 
 	windowPairs := make([]*externalapi.TrustedDataDataDAABlock, len(window))
 	for i, daaBlock := range window {
-		header, err := pm.blockHeaderStore.BlockHeader(pm.databaseContext, stagingArea, daaBlock.Hash)
+		daaDomainBlock, err := pm.blocksStore.Block(pm.databaseContext, stagingArea, daaBlock.Hash)
 		if err != nil {
 			return nil, err
 		}
 
 		windowPairs[i] = &externalapi.TrustedDataDataDAABlock{
-			Header:       header,
+			Block:        daaDomainBlock,
 			GHOSTDAGData: daaBlock.GHOSTDAGData,
 		}
 	}
